Add Segmenter tools
Adds a number of segmenter class objects that divide a vector<double>
object in pieces, based on value or number of elements.

Note that the segmenter produces 'VectorBlocks' which contain fixed
references to points inside the vector. This means that if the vector is
modified after the segmenter is initialized, the blocks are no longer
guaranteed to be valid. This behavior is not safe - so do not use this
object unless you know what you are doing. An example of proper usage
can be found in the test implementation. This commit includes:

- A test program

- A build rule in the Makefile