Add smart_string_compare()
Adds a smart_string_compare() function to the StringTools. This
comparison function takes two std::string arguments and proceeds to
compare the first to the second argument using auto-complete.

If the first string matches the leading substring of the second string,
the function returns 0. Otherwise it returns a positive or negative
number, depending on the type of comparison failure.

A test case module has also been added to the string test code.